restruct paginator tests
